+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Throw more actionable error objects than `FormatException` for errors. Errors
   were previously only correctly parsed in `generateContent` calls.
 - Add support for tuned models.
+- Add support for `batchEmbedContents` calls.
 
 ## 0.2.2